remove application templates provider from app.tsx
