--- .changeset/lucky-timers-argue.md
@@ -1,6 +1,0 @@
----
-"@wso2is/admin.template-core.v1": patch
-"@wso2is/console": patch
----
-
-Handle application templates provider based on available scopes
