+++ apps/console/src/app.tsx
@@ -17,13 +17,7 @@
  */
 
 import { BasicUserInfo, DecodedIDTokenPayload, useAuthContext } from "@asgardeo/auth-react";
-import {
-    AccessControlProvider,
-    AllFeatureInterface,
-    FeatureAccessConfigInterface,
-    FeatureGateInterface
-} from "@wso2is/access-control";
-import { ApplicationTemplateConstants } from "@wso2is/admin.application-templates.v1/constants/templates";
+import { AccessControlProvider, AllFeatureInterface, FeatureGateInterface } from "@wso2is/access-control";
 import { EventPublisher, PreLoader } from "@wso2is/admin.core.v1";
 import { ProtectedRoute } from "@wso2is/admin.core.v1/components";
 import { Config, DocumentationLinks } from "@wso2is/admin.core.v1/configs";
@@ -40,12 +34,9 @@ import { AppState } from "@wso2is/admin.core.v1/store";
 import { commonConfig } from "@wso2is/admin.extensions.v1";
 import { featureGateConfig } from "@wso2is/admin.extensions.v1/configs/feature-gate";
 import useGetAllFeatures from "@wso2is/admin.feature-gate.v1/api/use-get-all-features";
-import { ResourceTypes } from "@wso2is/admin.template-core.v1/models/templates";
-import ExtensionTemplatesProvider from "@wso2is/admin.template-core.v1/provider/extension-templates-provider";
 import { AppConstants as CommonAppConstants } from "@wso2is/core/constants";
 import { IdentityAppsApiException } from "@wso2is/core/exceptions";
-// eslint-disable-next-line no-restricted-imports
-import { CommonHelpers, hasRequiredScopes, isPortalAccessGranted } from "@wso2is/core/helpers";
+import { CommonHelpers, isPortalAccessGranted } from "@wso2is/core/helpers";
 import { RouteInterface, StorageIdentityAppsSettingsInterface, emptyIdentityAppsSettings } from "@wso2is/core/models";
 import { setI18nConfigs, setServiceResourceEndpoints } from "@wso2is/core/store";
 import { AuthenticateUtils, LocalStorageUtils } from "@wso2is/core/utils";
@@ -63,7 +54,7 @@ import has from "lodash-es/has";
 import isEmpty from "lodash-es/isEmpty";
 import set from "lodash-es/set";
 import * as moment from "moment";
-import React, { FunctionComponent, ReactElement, Suspense, useEffect, useMemo, useState } from "react";
+import React, { FunctionComponent, ReactElement, Suspense, useEffect, useState } from "react";
 import { Helmet } from "react-helmet";
 import { Trans } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
@@ -110,16 +101,6 @@ export const App: FunctionComponent<Record<string, never>> = (): ReactElement =>
         data: allFeatures,
         error: featureGateAPIException
     } = useGetAllFeatures();
-
-    const applicationsFeatureConfig: FeatureAccessConfigInterface = useSelector(
-        (state: AppState) => state?.config?.ui?.features?.applications
-    );
-
-    const hasApplicationTemplateViewPermissions: boolean = useMemo(() => hasRequiredScopes(
-        applicationsFeatureConfig,
-        applicationsFeatureConfig?.scopes?.read,
-        allowedScopes
-    ), [ applicationsFeatureConfig, allowedScopes ]);
 
     /**
      * Set the deployment configs in redux state.
@@ -486,52 +467,46 @@ export const App: FunctionComponent<Record<string, never>> = (): ReactElement =>
                                                 </Trans>)
                                             }
                                         />
-                                        <ExtensionTemplatesProvider
-                                            shouldFetch={ hasApplicationTemplateViewPermissions }
-                                            resourceType={ ResourceTypes.APPLICATIONS }
-                                            categories={ ApplicationTemplateConstants.SUPPORTED_CATEGORIES_INFO }
-                                        >
-                                            <Switch>
-                                                <Redirect
-                                                    exact
-                                                    from="/"
-                                                    to={ AppConstants.getAppHomePath() }
-                                                />
-                                                {
-                                                    baseRoutes.map((route: RouteInterface, index: number) => {
-                                                        return (
-                                                            route.protected ?
-                                                                (
-                                                                    <ProtectedRoute
-                                                                        component={ route.component }
-                                                                        path={ route.path }
-                                                                        key={ index }
-                                                                        exact={ route.exact }
-                                                                    />
-                                                                )
-                                                                :
-                                                                (
-                                                                    <Route
-                                                                        path={ route.path }
-                                                                        render={
-                                                                            (props:  RouteComponentProps<
-                                                                                { [p: string]: string },
-                                                                                StaticContext, unknown
-                                                                            >) => {
-                                                                                return (<route.component
-                                                                                    { ...props }
-                                                                                />);
-                                                                            }
+                                        <Switch>
+                                            <Redirect
+                                                exact
+                                                from="/"
+                                                to={ AppConstants.getAppHomePath() }
+                                            />
+                                            {
+                                                baseRoutes.map((route: RouteInterface, index: number) => {
+                                                    return (
+                                                        route.protected ?
+                                                            (
+                                                                <ProtectedRoute
+                                                                    component={ route.component }
+                                                                    path={ route.path }
+                                                                    key={ index }
+                                                                    exact={ route.exact }
+                                                                />
+                                                            )
+                                                            :
+                                                            (
+                                                                <Route
+                                                                    path={ route.path }
+                                                                    render={
+                                                                        (props:  RouteComponentProps<
+                                                                            { [p: string]: string },
+                                                                            StaticContext, unknown
+                                                                        >) => {
+                                                                            return (<route.component
+                                                                                { ...props }
+                                                                            />);
                                                                         }
-                                                                        key={ index }
-                                                                        exact={ route.exact }
-                                                                    />
-                                                                )
-                                                        );
-                                                    })
-                                                }
-                                            </Switch>
-                                        </ExtensionTemplatesProvider>
+                                                                    }
+                                                                    key={ index }
+                                                                    exact={ route.exact }
+                                                                />
+                                                            )
+                                                    );
+                                                })
+                                            }
+                                        </Switch>
                                     </>
                                 </SessionManagementProvider>
                             </AccessControlProvider>
